HOTT-1973 Refactor Step2: move the retries method into a module called Retriable.
