+++ app/services/beta/search/perform_search_service.rb
@@ -1,15 +1,18 @@
 require 'tariff_jsonapi_parser'
 require 'errors'
+require 'retriable'
 
 module Beta
   module Search
     class PerformSearchService
+      include Retriable
+
       def initialize(query, filters = {})
         @query = query
         @filters = filters
       end
 
-      def call()
+      def call
         retries(Faraday::Error, UnparseableResponseError) do
           path = "/api/beta/search?#{query_params}"
           response = api.get(path)
@@ -44,18 +47,6 @@ module Beta
 
       def api
         TradeTariffFrontend::ServiceChooser.api_client
-      end
-
-      def retries(*rescue_errors)
-        retries ||= 0
-        yield if block_given?
-      rescue *rescue_errors
-        if retries < Rails.configuration.x.http.max_retry
-          retries += 1
-          retry
-        else
-          raise
-        end
       end
     end
   end
