+++ lib/api_entity.rb
@@ -3,6 +3,7 @@ require 'multi_json'
 require 'active_model'
 require 'tariff_jsonapi_parser'
 require 'errors'
+require 'retriable'
 
 module ApiEntity
   extend ActiveSupport::Concern
@@ -85,6 +86,8 @@ private
   end
 
   module ClassMethods
+    include Retriable
+
     delegate :get, :post, to: :api
 
     def relationships
@@ -95,18 +98,6 @@ private
       retries(Faraday::Error, UnparseableResponseError) do
         resp = api.get("/#{name.pluralize.underscore}/#{id}", opts)
         new parse_jsonapi(resp)
-      end
-    end
-
-    def retries(*rescue_errors)
-      retries ||= 0
-      yield if block_given?
-    rescue *rescue_errors
-      if retries < Rails.configuration.x.http.max_retry
-        retries += 1
-        retry
-      else
-        raise
       end
     end
